+++ tests/ui/c/inc/t_ui_tearing.h
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2019 MicroEJ Corp. All rights reserved.
+ * For demonstration purpose only.
+ * MicroEJ Corp. PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
+ */
+
+/* Prevent recursive inclusion */
+
+#ifndef __T_UI_TEARING_H
+#define __T_UI_TEARING_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/* Public function declarations */
+
+/**
+ * @brief XXX
+ */
+TestRef T_UI_TEARING_tests(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif